Super broken. But, there is some forward progress!
Need to understand XML fields that contain ':'. Go does not like
parsing them, but we need to be able to serialize fields that
follow such a pattern (does not currently work).

Initial work on replacing sections sorta works, but that is
plagued by the problem described above.

I also abstracted how hardware fields can be modified. I think
the idea of injecting a function seems to work well.

Still a lot that needs to be done. Tests are borked.