Fixed broken ovf editing logic.
This was very noticable when making multiple changes because the
adjustments were not applied properly, and getting information
about the original line data was done using the editing data.

This is worked around by maintaining a copy of the original data.
It is not ideal, but it will get us by for now.